Refactor to single code path for Refs
