+++ packages/ui/react-ui-form/src/util/index.ts
@@ -1,7 +1,8 @@
 //
-// Copyright 2024 DXOS.org
+// Copyright 2025 DXOS.org
 //
 
-export * from './schema';
 export * from './field';
 export * from './formatting';
+export * from './refs';
+export * from './schema';
